Refactor component base class
